+++ templates/accesoLogin.html
@@ -112,6 +112,8 @@
     <!-- Bootstrap JS -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/5.3.0/js/bootstrap.bundle.min.js"></script>
     <!-- Custom JS -->
-    <script src="../static/js/scriptLogin.js"></script>
+    <script src="../static/js/Login/FacialRecognitionLogin.js"></script>
+    <script src="../static/js/Login/TraditionalLogin.js"></script>
+    <script src="../static/js/Login/app.js"></script></scriptsrc>
 </body>
 </html>
